Conflicts:
	.idea/mg4j-nyu.iml
	build.xml
	classes/production/mg4j-nyu/edu/nyu/tandon/experiments/ListSpeed.class
	classes/production/mg4j-nyu/edu/nyu/tandon/experiments/PrunedQuery.class
	classes/production/mg4j-nyu/edu/nyu/tandon/experiments/RandomAccess.class
	classes/production/mg4j-nyu/edu/nyu/tandon/experiments/RawDocHits$1.class
	classes/production/mg4j-nyu/edu/nyu/tandon/experiments/RawDocHits$Command.class
	classes/production/mg4j-nyu/edu/nyu/tandon/experiments/RawDocHits$OutputType.class
	classes/production/mg4j-nyu/edu/nyu/tandon/experiments/RawDocHits.class
	classes/production/mg4j-nyu/edu/nyu/tandon/query/PrunedQueryEngine.class
	classes/production/mg4j-nyu/edu/nyu/tandon/query/Query$1.class
	classes/production/mg4j-nyu/edu/nyu/tandon/query/Query$2.class
	classes/production/mg4j-nyu/edu/nyu/tandon/query/Query$Command.class
	classes/production/mg4j-nyu/edu/nyu/tandon/query/Query$OutputType.class
	classes/production/mg4j-nyu/edu/nyu/tandon/query/Query.class
	classes/production/mg4j-nyu/edu/nyu/tandon/query/QueryEngine.class
	classes/production/mg4j-nyu/edu/nyu/tandon/tool/xdoc/XDOCConcatenate.class
	classes/production/mg4j-nyu/edu/nyu/tandon/tool/xdoc/XDOCScan$1.class
	classes/production/mg4j-nyu/edu/nyu/tandon/tool/xdoc/XDOCScan.class
	classes/production/mg4j-nyu/edu/nyu/tandon/tool/xdoc/printXDOC.class
	classes/production/mg4j-nyu/edu/nyu/tandon/utils/languageModel.class
	classes/production/mg4j-nyu/it/unimi/di/big/mg4j/document/WarcDocumentSequence_NYU$1.class
	classes/production/mg4j-nyu/it/unimi/di/big/mg4j/document/WarcDocumentSequence_NYU.class
	classes/production/mg4j-nyu/it/unimi/di/big/mg4j/tool/Scan$1.class
	classes/production/mg4j-nyu/it/unimi/di/big/mg4j/tool/Scan$2.class
	classes/production/mg4j-nyu/it/unimi/di/big/mg4j/tool/Scan$3.class
	classes/production/mg4j-nyu/it/unimi/di/big/mg4j/tool/Scan$4.class
	classes/production/mg4j-nyu/it/unimi/di/big/mg4j/tool/Scan$5.class
	classes/production/mg4j-nyu/it/unimi/di/big/mg4j/tool/Scan$Completeness.class
	classes/production/mg4j-nyu/it/unimi/di/big/mg4j/tool/Scan$IndexingType.class
	classes/production/mg4j-nyu/it/unimi/di/big/mg4j/tool/Scan$PayloadAccumulator$1.class
	classes/production/mg4j-nyu/it/unimi/di/big/mg4j/tool/Scan$PayloadAccumulator.class
	classes/production/mg4j-nyu/it/unimi/di/big/mg4j/tool/Scan.class
	classes/production/mg4j-nyu/it/unimi/di/law/warc/io/AbstractWarcReader_NYU.class
	classes/production/mg4j-nyu/it/unimi/di/law/warc/io/UncompressedWarcReader_NYU.class
	lib/activation.jar
	lib/antlr.jar
	lib/apache-mime4j-core.jar
	lib/apache-mime4j-dom.jar
	lib/asm.jar
	lib/aspectjrt.jar
	lib/avro-ipc.jar
	lib/avro.jar
	lib/bcmail-jdk15on.jar
	lib/bcpkix-jdk15on.jar
	lib/bcprov-jdk15on.jar
	lib/bliki-core.jar
	lib/boilerpipe.jar
	lib/bzip2.jar
	lib/c3p0.jar
	lib/cdm.jar
	lib/collections-generic.jar
	lib/colt.jar
	lib/commons-beanutils.jar
	lib/commons-cli.jar
	lib/commons-codec.jar
	lib/commons-collections.jar
	lib/commons-compress.jar
	lib/commons-configuration.jar
	lib/commons-csv.jar
	lib/commons-digester.jar
	lib/commons-el.jar
	lib/commons-exec.jar
	lib/commons-fileupload.jar
	lib/commons-httpclient.jar
	lib/commons-io.jar
	lib/commons-lang.jar
	lib/commons-lang3.jar
	lib/commons-logging-api.jar
	lib/commons-logging.jar
	lib/commons-math.jar
	lib/commons-math3.jar
	lib/commons-net.jar
	lib/commons-validator.jar
	lib/commons-vfs2.jar
	lib/concurrent.jar
	lib/core.jar
	lib/cxf-core.jar
	lib/cxf-rt-frontend-jaxrs.jar
	lib/cxf-rt-rs-client.jar
	lib/cxf-rt-transports-http.jar
	lib/dnsjava.jar
	lib/dsiutils.jar
	lib/ehcache-core.jar
	lib/fastutil.jar
	lib/fontbox.jar
	lib/geoapi.jar
	lib/grib.jar
	lib/guava.jar
	lib/hadoop-annotations.jar
	lib/hadoop-auth.jar
	lib/hadoop-common.jar
	lib/hamcrest-core.jar
	lib/httpclient.jar
	lib/httpcore.jar
	lib/httpmime.jar
	lib/httpservices.jar
	lib/isoparser.jar
	lib/jackcess-encrypt.jar
	lib/jackcess.jar
	lib/jackson-core-asl.jar
	lib/jackson-jaxrs.jar
	lib/jackson-mapper-asl.jar
	lib/jackson-xc.jar
	lib/jai4j.jar
	lib/jasper-compiler.jar
	lib/jasper-runtime.jar
	lib/java-libpst.jar
	lib/javacc.jar
	lib/javax.annotation-api.jar
	lib/javax.ws.rs-api.jar
	lib/jaxb-api.jar
	lib/jaxb-impl.jar
	lib/jcip-annotations.jar
	lib/jcl-over-slf4j.jar
	lib/jcommander.jar
	lib/jdiff.jar
	lib/jdom.jar
	lib/jdom2.jar
	lib/je.jar
	lib/jempbox.jar
	lib/jericho-html-dev.jar
	lib/jersey-core.jar
	lib/jersey-json.jar
	lib/jersey-server.jar
	lib/jets3t.jar
	lib/jettison.jar
	lib/jetty-util.jar
	lib/jetty.jar
	lib/jgroups.jar
	lib/jhighlight.jar
	lib/jj2000.jar
	lib/jmatio.jar
	lib/jna.jar
	lib/joda-time.jar
	lib/jsap.jar
	lib/json-simple.jar
	lib/json.jar
	lib/jsoup.jar
	lib/jsp-api.jar
	lib/jsr-275.jar
	lib/jung-algorithms.jar
	lib/jung-api.jar
	lib/jung-io.jar
	lib/junit.jar
	lib/juniversalchardet.jar
	lib/junrar.jar
	lib/jwnl.jar
	lib/kfs.jar
	lib/log4j-over-slf4j.jar
	lib/logback-classic.jar
	lib/logback-core.jar
	lib/mail.jar
	lib/maven-scm-api.jar
	lib/maven-scm-provider-svn-commons.jar
	lib/maven-scm-provider-svnexe.jar
	lib/metadata-extractor.jar
	lib/mx4j-tools.jar
	lib/mx4j.jar
	lib/mysql-connector-java.jar
	lib/netcdf4.jar
	lib/netty.jar
	lib/opennlp-maxent.jar
	lib/opennlp-tools.jar
	lib/oro.jar
	lib/paranamer.jar
	lib/pdfbox.jar
	lib/plexus-utils.jar
	lib/poi-ooxml-schemas.jar
	lib/poi-ooxml.jar
	lib/poi-scratchpad.jar
	lib/poi.jar
	lib/pojo-mbean.jar
	lib/protobuf-java.jar
	lib/quartz.jar
	lib/regexp.jar
	lib/rome.jar
	lib/servlet-api.jar
	lib/sis-metadata.jar
	lib/sis-netcdf.jar
	lib/sis-referencing.jar
	lib/sis-storage.jar
	lib/sis-utility.jar
	lib/slf4j-api.jar
	lib/snappy-java.jar
	lib/sslext.jar
	lib/stax-api.jar
	lib/stax2-api.jar
	lib/struts.jar
	lib/sux4j.jar
	lib/tagsoup.jar
	lib/tika-core.jar
	lib/tika-parsers.jar
	lib/udunits.jar
	lib/umlgraph.jar
	lib/velocity-dep.jar
	lib/velocity-tools.jar
	lib/velocity.jar
	lib/vorbis-java-core.jar
	lib/vorbis-java-tika.jar
	lib/webgraph-big.jar
	lib/webgraph.jar
	lib/woodstox-core-asl.jar
	lib/wstx-asl.jar
	lib/xalan.jar
	lib/xmlbeans.jar
	lib/xmlenc.jar
	lib/xmlschema-core.jar
	lib/xmpcore.jar
	lib/xz.jar
	scripts/gov2-docHits.sh
	scripts/logback.xml
	src/edu/nyu/tandon/experiments/ListSpeed.java
	src/edu/nyu/tandon/experiments/PrunedQuery.java
	src/edu/nyu/tandon/experiments/RandomAccess.java
	src/edu/nyu/tandon/experiments/RawDocHits.java
	src/edu/nyu/tandon/query/PrunedQueryEngine.java
	src/edu/nyu/tandon/query/Query.java
	src/edu/nyu/tandon/query/QueryEngine.java
	src/edu/nyu/tandon/tool/xdoc/XDOCConcatenate.java
	src/edu/nyu/tandon/tool/xdoc/XDOCScan.java
	src/edu/nyu/tandon/utils/languageModel.java
	src/it/unimi/di/big/mg4j/document/WarcDocumentSequence_NYU.java
	src/it/unimi/di/big/mg4j/tool/Scan.java
	src/it/unimi/di/law/warc/io/AbstractWarcReader_NYU.java
	src/it/unimi/di/law/warc/io/UncompressedWarcReader_NYU.java

+++ src/edu/nyu/tandon/experiments/RandomAccess.java
@@ -3,7 +3,6 @@ package edu.nyu.tandon.experiments;
 import com.martiansoftware.jsap.*;
 import edu.nyu.tandon.query.Query;
 import edu.nyu.tandon.query.QueryEngine;
-import edu.nyu.tandon.utils.*;
 import it.unimi.di.big.mg4j.index.Index;
 import it.unimi.di.big.mg4j.index.IndexIterator;
 import it.unimi.di.big.mg4j.index.TermProcessor;
@@ -27,24 +26,24 @@ import static it.unimi.di.big.mg4j.search.DocumentIterator.END_OF_LIST;
 public class RandomAccess {
 
     @SuppressWarnings("unchecked")
-    public static void main( final String[] arg ) throws Exception {
+    public static void main(final String[] arg) throws Exception {
 
         SimpleJSAP jsap = new SimpleJSAP(
-            ListSpeed.class.getName(),
-            "Loads random list, primes, and measure random posting access times",
-            new Parameter[] {
-                new UnflaggedOption( "basenameWeight", JSAP.STRING_PARSER, JSAP.NO_DEFAULT, JSAP.REQUIRED, JSAP.GREEDY, "The indices that the servlet will use. Indices are specified using their basename, optionally followed by a colon and a double representing the weight used to score results from that index. Indices without a specified weight are weighted 1." ),
-                new FlaggedOption( "random", JSAP.INTEGER_PARSER, "1000", JSAP.NOT_REQUIRED, 'r', "random", "The # of postings to measure" ),
-                new FlaggedOption( "threshold", JSAP.INTEGER_PARSER, "100000", JSAP.NOT_REQUIRED, 't', "threshold", "The min # terms in a document" ),
-                new FlaggedOption( "samples", JSAP.INTEGER_PARSER, "1", JSAP.NOT_REQUIRED, 's', "samples", "The # of terms to measure" )
-        });
+                ListSpeed.class.getName(),
+                "Loads random list, primes, and measure random posting access times",
+                new Parameter[]{
+                        new UnflaggedOption("basenameWeight", JSAP.STRING_PARSER, JSAP.NO_DEFAULT, JSAP.REQUIRED, JSAP.GREEDY, "The indices that the servlet will use. Indices are specified using their basename, optionally followed by a colon and a double representing the weight used to score results from that index. Indices without a specified weight are weighted 1."),
+                        new FlaggedOption("random", JSAP.INTEGER_PARSER, "1000", JSAP.NOT_REQUIRED, 'r', "random", "The # of postings to measure"),
+                        new FlaggedOption("threshold", JSAP.INTEGER_PARSER, "100000", JSAP.NOT_REQUIRED, 't', "threshold", "The min # terms in a document"),
+                        new FlaggedOption("samples", JSAP.INTEGER_PARSER, "1", JSAP.NOT_REQUIRED, 's', "samples", "The # of terms to measure")
+                });
 
-        final JSAPResult jsapResult = jsap.parse( arg );
-        if ( jsap.messagePrinted() ) return;
+        final JSAPResult jsapResult = jsap.parse(arg);
+        if (jsap.messagePrinted()) return;
 
-        final String[] basenameWeight = jsapResult.getStringArray( "basenameWeight" );
+        final String[] basenameWeight = jsapResult.getStringArray("basenameWeight");
 
-        final Object2ReferenceLinkedOpenHashMap<String,Index> indexMap = new Object2ReferenceLinkedOpenHashMap<String,Index>( Hash.DEFAULT_INITIAL_SIZE, .5f );
+        final Object2ReferenceLinkedOpenHashMap<String, Index> indexMap = new Object2ReferenceLinkedOpenHashMap<String, Index>(Hash.DEFAULT_INITIAL_SIZE, .5f);
         final Reference2DoubleOpenHashMap<Index> index2Weight = new Reference2DoubleOpenHashMap<Index>();
 
         // load index
@@ -54,48 +53,48 @@ public class RandomAccess {
         // get stats
         final long numberOfDocuments = indexMap.values().iterator().next().numberOfDocuments;
         final long numberOfTerms = indexMap.values().iterator().next().numberOfTerms;
-        final long numberOfSamples = jsapResult.getInt("samples", (int)numberOfDocuments);
-        final long numberOPostings = jsapResult.getInt("random", (int)1000);
+        final long numberOfSamples = jsapResult.getInt("samples", (int) numberOfDocuments);
+        final long numberOPostings = jsapResult.getInt("random", 1000);
 
-        final Object2ObjectOpenHashMap<String,TermProcessor> termProcessors = new Object2ObjectOpenHashMap<String,TermProcessor>( indexMap.size() );
-        for( String alias: indexMap.keySet() ) termProcessors.put( alias, indexMap.get( alias ).termProcessor );
-        final SimpleParser simpleParser = new SimpleParser( indexMap.keySet(), indexMap.firstKey(), termProcessors );
+        final Object2ObjectOpenHashMap<String, TermProcessor> termProcessors = new Object2ObjectOpenHashMap<String, TermProcessor>(indexMap.size());
+        for (String alias : indexMap.keySet()) termProcessors.put(alias, indexMap.get(alias).termProcessor);
+        final SimpleParser simpleParser = new SimpleParser(indexMap.keySet(), indexMap.firstKey(), termProcessors);
         final Reference2ReferenceMap<Index, Object> index2Parser = new Reference2ReferenceOpenHashMap<Index, Object>();
-        final QueryEngine queryEngine = new QueryEngine( simpleParser, new DocumentIteratorBuilderVisitor( indexMap, index2Parser, indexMap.get( indexMap.firstKey() ), 2048 ), indexMap );
-        queryEngine.setWeights( index2Weight );
+        final QueryEngine queryEngine = new QueryEngine(simpleParser, new DocumentIteratorBuilderVisitor(indexMap, index2Parser, indexMap.get(indexMap.firstKey()), 2048), indexMap);
+        queryEngine.setWeights(index2Weight);
 
-        queryEngine.score( new Scorer[] {}, new double[] { 1, 1 } );
+        queryEngine.score(new Scorer[]{}, new double[]{1, 1});
         queryEngine.intervalSelector = new IntervalSelector();
         queryEngine.multiplex = false;
 
-        Query query = new Query( queryEngine );
+        Query query = new Query(queryEngine);
         query.displayMode = Query.OutputType.TIME;
-        query.maxOutput = jsapResult.getInt("results", (int)numberOfDocuments);
+        query.maxOutput = jsapResult.getInt("results", (int) numberOfDocuments);
         int threshold = jsapResult.getInt("threshold", 100000);
         query.interpretCommand("$mode time");
 
-        System.err.println( "MG4J experiment - posting access times." );
+        System.err.println("MG4J experiment - posting access times.");
 
         // seed with random documents
         Set<Integer> set = new TreeSet<Integer>();
-        for (int i=0; i<numberOPostings; i++) {
-            set.add(rand(0, (int)numberOfDocuments-1));
+        for (int i = 0; i < numberOPostings; i++) {
+            set.add(rand(0, (int) numberOfDocuments - 1));
             System.err.print("\rAdding random postings: " + i);
         }
         Integer[] docs = set.toArray(new Integer[0]);
 
         // find 10 terms with at least t postings
         long[] terms = new long[10];
-        int numTerms  = 0;
-        for (int i=0; i< numberOfTerms && numTerms < 10; i++) {
+        int numTerms = 0;
+        for (int i = 0; i < numberOfTerms && numTerms < 10; i++) {
             System.err.print("\rSearching terms: " + i);
-            IndexIterator list =  indexMap.values().iterator().next().documents(i);
+            IndexIterator list = indexMap.values().iterator().next().documents(i);
             list.dispose();
             if (list.frequency() >= threshold) {
                 terms[numTerms++] = list.termNumber();
             }
         }
-        if (numTerms==0) return;
+        if (numTerms == 0) return;
 
         System.err.println();
         double totSequential = 0;
@@ -109,12 +108,12 @@ public class RandomAccess {
         long document = 0;
         long count = 0;
 
-        for (int k=0; k<numTerms; k++) {
+        for (int k = 0; k < numTerms; k++) {
 
             try {
 
                 // get pointer to the term list
-                IndexIterator list =  indexMap.values().iterator().next().documents(terms[k]);
+                IndexIterator list = indexMap.values().iterator().next().documents(terms[k]);
                 if (list.frequency() == 0) return;
                 totSeqPostings += list.frequency();
                 list.dispose();
@@ -122,38 +121,36 @@ public class RandomAccess {
                 double total = 0;
 
                 // sequential access metrics; first pass to prime the system
-                for (int i=0; i<11; i++)
-                {
-                    list =  indexMap.values().iterator().next().documents(terms[k]);
+                for (int i = 0; i < 11; i++) {
+                    list = indexMap.values().iterator().next().documents(terms[k]);
                     long time = -System.nanoTime();
                     // scan entire list
-                    while ((document = list.nextDocument() ) != END_OF_LIST );
+                    while ((document = list.nextDocument()) != END_OF_LIST) ;
                     time += System.nanoTime();
-                    if (i>0) total += time;
+                    if (i > 0) total += time;
                     list.dispose();
                 }
                 total /= 10;
                 totSequential += total;
                 System.out.println("Term " + k + ": " + list.frequency() + " postings");
-                System.out.println("Sequential access:  " + total / 1000000. + " ms,  " + Util.format((list.frequency() * 1000000000.0) / total) + " postings/s,   " + Util.format( total / (double)list.frequency() ) + " ns/posting" );
+                System.out.println("Sequential access:  " + total / 1000000. + " ms,  " + Util.format((list.frequency() * 1000000000.0) / total) + " postings/s,   " + Util.format(total / (double) list.frequency()) + " ns/posting");
 
                 // random access metrics
                 total = 0;
                 long hits = 0;
                 long misses = 0;
                 totRandPostings += docs.length;
-                for (int i=0; i<11; i++) {
+                for (int i = 0; i < 11; i++) {
 
-                    list =  indexMap.values().iterator().next().documents(terms[k]);
+                    list = indexMap.values().iterator().next().documents(terms[k]);
                     document = -1;
-                    if (i>0) {
+                    if (i > 0) {
                         long time = -System.nanoTime();
-                        for (int j=0; j<docs.length && document != END_OF_LIST; j++)
+                        for (int j = 0; j < docs.length && document != END_OF_LIST; j++)
                             document = list.skipTo(docs[j]);
                         time += System.nanoTime();
                         total += time;
-                    }
-                    else {
+                    } else {
                         for (int j = 0; j < docs.length && document != END_OF_LIST; j++) {
                             document = list.skipTo(docs[j]);
                             if (document == docs[j]) hits++;
@@ -167,10 +164,9 @@ public class RandomAccess {
                 totHits += hits;
                 totMisses += misses;
 
-                System.out.println("Random access:  " + docs.length + " lookups,  " + hits + " hits,  " + misses + " misses,  " + total / 1000000. + " ms,  " + Util.format((docs.length * 1000000000.0) / total) + " postings/s, " + Util.format( total / (double)docs.length ) + " ns/posting" );
+                System.out.println("Random access:  " + docs.length + " lookups,  " + hits + " hits,  " + misses + " misses,  " + total / 1000000. + " ms,  " + Util.format((docs.length * 1000000000.0) / total) + " postings/s, " + Util.format(total / (double) docs.length) + " ns/posting");
                 System.out.println();
-            }
-            catch( Exception e ) {
+            } catch (Exception e) {
                 System.err.println(e);
             }
         }
@@ -184,9 +180,9 @@ public class RandomAccess {
 
         System.out.println();
         System.out.println();
-        System.out.println("Average: " + + totSeqPostings + " postings");
-        System.out.println("Sequential access: "  + totSequential / 1000000. + " ms; " + Util.format((totSeqPostings * 1000000000.0) / totSequential) + " postings/s, " + Util.format( totSequential / (double)totSeqPostings ) + " ns/posting" );
-        System.out.println("Random access: " + docs.length+ " lookups,  " + totHits + " hits,  " + totMisses + " misses,  " + totRandom / 1000000. + " ms,  " + Util.format((totRandPostings * 1000000000.0) / totRandom) + " postings/s, " + Util.format( totRandom / (double)docs.length ) + " ns/posting" );
+        System.out.println("Average: " + +totSeqPostings + " postings");
+        System.out.println("Sequential access: " + totSequential / 1000000. + " ms; " + Util.format((totSeqPostings * 1000000000.0) / totSequential) + " postings/s, " + Util.format(totSequential / (double) totSeqPostings) + " ns/posting");
+        System.out.println("Random access: " + docs.length + " lookups,  " + totHits + " hits,  " + totMisses + " misses,  " + totRandom / 1000000. + " ms,  " + Util.format((totRandPostings * 1000000000.0) / totRandom) + " postings/s, " + Util.format(totRandom / (double) docs.length) + " ns/posting");
 
     }
